dirty fix an invalid form control is not focusable error
Chrome wants to focus on a control that is required
but still empty so that it can pop up the message
'Please fill out this field'. However, if the control is hidden
at the point that Chrome wants to pop up the message,
that is at the time of form submission,
Chrome can't focus on the control because it is hidden,
therefore the form won't submit.
So, to get around the problem, when a control is hidden by javascript,
we must remove the 'required' attribute from that control or
add novalidate to the form.