make NoteSerializer writable
If you try to use a writable nested serializer without writing a custom
create() and/or update() method you'll see an assertion
error when you attempt to save the serializer.
To use writable nested serialization we need to declare a nested
field on the serializer class,
and write the create() and/or update() methods explicitly.